Move configuration related classes into sub directory
