+++ src/main/java/sonia/scm/jira/JiraIssueTrackerProvider.java
@@ -27,6 +27,9 @@ package sonia.scm.jira;
 import sonia.scm.issuetracker.api.IssueTracker;
 import sonia.scm.issuetracker.spi.IssueTrackerBuilder;
 import sonia.scm.issuetracker.spi.IssueTrackerProvider;
+import sonia.scm.jira.config.JiraConfiguration;
+import sonia.scm.jira.config.JiraConfigurationResolver;
+import sonia.scm.jira.config.JiraStateChanger;
 import sonia.scm.jira.rest.RestApi;
 import sonia.scm.net.ahc.AdvancedHttpClient;
 import sonia.scm.repository.Repository;
