--- src/main/java/sonia/scm/jira/RepositoryHalEnricher.java
@@ -1,64 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020-present Cloudogu GmbH and Contributors
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package sonia.scm.jira;
-
-import sonia.scm.api.v2.resources.Enrich;
-import sonia.scm.api.v2.resources.HalAppender;
-import sonia.scm.api.v2.resources.LinkBuilder;
-import sonia.scm.api.v2.resources.HalEnricher;
-import sonia.scm.api.v2.resources.HalEnricherContext;
-import sonia.scm.api.v2.resources.ScmPathInfoStore;
-import sonia.scm.plugin.Extension;
-import sonia.scm.repository.Repository;
-
-import javax.inject.Inject;
-import javax.inject.Provider;
-
-@Extension
-@Enrich(Repository.class)
-public class RepositoryHalEnricher implements HalEnricher {
-
-  private final Provider<ScmPathInfoStore> scmPathInfoStore;
-  private final JiraGlobalContext jiraContext;
-  private final JiraPermissions permissions;
-
-  @Inject
-  public RepositoryHalEnricher(Provider<ScmPathInfoStore> scmPathInfoStore, JiraGlobalContext jiraContext, JiraPermissions permissions) {
-    this.scmPathInfoStore = scmPathInfoStore;
-    this.jiraContext = jiraContext;
-    this.permissions = permissions;
-  }
-
-  @Override
-  public void enrich(HalEnricherContext context, HalAppender appender) {
-    Repository repository = context.oneRequireByType(Repository.class);
-    if (!jiraContext.getGlobalConfiguration().isDisableRepositoryConfiguration() && permissions.isPermittedReadRepositoryConfig(repository)) {
-      String globalJiraConfigUrl = new LinkBuilder(scmPathInfoStore.get().get(), JiraConfigurationResource.class)
-        .method("getForRepository")
-        .parameters(repository.getNamespace(), repository.getName())
-        .href();
-      appender.appendLink("jiraConfig", globalJiraConfigUrl);
-    }
-  }
-}
