+++ src/main/java/sonia/scm/jira/commitmessagechecker/JiraCommitMessageIssueKeyValidator.java
@@ -35,9 +35,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import sonia.scm.ContextEntry;
-import sonia.scm.jira.JiraConfiguration;
-import sonia.scm.jira.JiraConfigurationResolver;
-import sonia.scm.jira.JiraGlobalContext;
+import sonia.scm.jira.config.JiraConfiguration;
+import sonia.scm.jira.config.JiraConfigurationResolver;
 import sonia.scm.plugin.Extension;
 import sonia.scm.plugin.Requires;
 import sonia.scm.util.GlobUtil;
