+++ src/main/java/sonia/scm/jira/config/BaseMapper.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package sonia.scm.jira;
+package sonia.scm.jira.config;
 
 import org.apache.commons.lang.StringUtils;
 import org.mapstruct.AfterMapping;
