+++ src/main/java/sonia/scm/jira/config/IndexHalEnricher.java
@@ -21,32 +21,42 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+package sonia.scm.jira.config;
 
-package sonia.scm.jira;
-
-//~--- non-JDK imports --------------------------------------------------------
-
-import com.google.inject.AbstractModule;
-import org.mapstruct.factory.Mappers;
+import sonia.scm.api.v2.resources.Enrich;
+import sonia.scm.api.v2.resources.Index;
+import sonia.scm.api.v2.resources.HalAppender;
+import sonia.scm.api.v2.resources.LinkBuilder;
+import sonia.scm.api.v2.resources.HalEnricher;
+import sonia.scm.api.v2.resources.HalEnricherContext;
+import sonia.scm.api.v2.resources.ScmPathInfoStore;
+import sonia.scm.jira.JiraPermissions;
 import sonia.scm.plugin.Extension;
 
-/**
- *
- * @author Sebastian Sdorra
- */
-@Extension
-public class JiraModule extends AbstractModule
-{
+import javax.inject.Inject;
+import javax.inject.Provider;
 
-  /**
-   * Method description
-   *
-   */
+@Extension
+@Enrich(Index.class)
+public class IndexHalEnricher implements HalEnricher {
+
+  private final Provider<ScmPathInfoStore> scmPathInfoStore;
+  private final JiraPermissions permissions;
+
+  @Inject
+  public IndexHalEnricher(Provider<ScmPathInfoStore> scmPathInfoStore, JiraPermissions permissions) {
+    this.scmPathInfoStore = scmPathInfoStore;
+    this.permissions = permissions;
+  }
+
   @Override
-  protected void configure()
-  {
-    bind(JiraGlobalContext.class);
-    bind(JiraConfigurationMapper.class).to(Mappers.getMapper(JiraConfigurationMapper.class).getClass());
-    bind(JiraGlobalConfigurationMapper.class).to(Mappers.getMapper(JiraGlobalConfigurationMapper.class).getClass());
+  public void enrich(HalEnricherContext context, HalAppender appender) {
+    if (permissions.isPermittedReadGlobalConfig()) {
+      String globalJiraConfigUrl = new LinkBuilder(scmPathInfoStore.get().get(), JiraConfigurationResource.class)
+        .method("get")
+        .parameters()
+        .href();
+      appender.appendLink("jiraConfig", globalJiraConfigUrl);
+    }
   }
 }
