+++ src/main/java/sonia/scm/jira/config/RepositoryHalEnricher.java
@@ -21,39 +21,43 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package sonia.scm.jira;
+package sonia.scm.jira.config;
 
 import sonia.scm.api.v2.resources.Enrich;
-import sonia.scm.api.v2.resources.Index;
 import sonia.scm.api.v2.resources.HalAppender;
 import sonia.scm.api.v2.resources.LinkBuilder;
 import sonia.scm.api.v2.resources.HalEnricher;
 import sonia.scm.api.v2.resources.HalEnricherContext;
 import sonia.scm.api.v2.resources.ScmPathInfoStore;
+import sonia.scm.jira.JiraPermissions;
 import sonia.scm.plugin.Extension;
+import sonia.scm.repository.Repository;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
 
 @Extension
-@Enrich(Index.class)
-public class IndexHalEnricher implements HalEnricher {
+@Enrich(Repository.class)
+public class RepositoryHalEnricher implements HalEnricher {
 
   private final Provider<ScmPathInfoStore> scmPathInfoStore;
+  private final JiraConfigurationStore jiraContext;
   private final JiraPermissions permissions;
 
   @Inject
-  public IndexHalEnricher(Provider<ScmPathInfoStore> scmPathInfoStore, JiraPermissions permissions) {
+  public RepositoryHalEnricher(Provider<ScmPathInfoStore> scmPathInfoStore, JiraConfigurationStore jiraContext, JiraPermissions permissions) {
     this.scmPathInfoStore = scmPathInfoStore;
+    this.jiraContext = jiraContext;
     this.permissions = permissions;
   }
 
   @Override
   public void enrich(HalEnricherContext context, HalAppender appender) {
-    if (permissions.isPermittedReadGlobalConfig()) {
+    Repository repository = context.oneRequireByType(Repository.class);
+    if (!jiraContext.getGlobalConfiguration().isDisableRepositoryConfiguration() && permissions.isPermittedReadRepositoryConfig(repository)) {
       String globalJiraConfigUrl = new LinkBuilder(scmPathInfoStore.get().get(), JiraConfigurationResource.class)
-        .method("get")
-        .parameters()
+        .method("getForRepository")
+        .parameters(repository.getNamespace(), repository.getName())
         .href();
       appender.appendLink("jiraConfig", globalJiraConfigUrl);
     }
