+++ src/main/java/sonia/scm/jira/update/JiraV2GlobalConfigUpdateStep.java
@@ -26,8 +26,8 @@ package sonia.scm.jira.update;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import sonia.scm.issuetracker.XmlEncryptionAdapter;
-import sonia.scm.jira.JiraGlobalConfiguration;
-import sonia.scm.jira.XmlStringMapAdapter;
+import sonia.scm.jira.config.JiraGlobalConfiguration;
+import sonia.scm.jira.config.XmlStringMapAdapter;
 import sonia.scm.migration.UpdateStep;
 import sonia.scm.plugin.Extension;
 import sonia.scm.store.ConfigurationStoreFactory;
