--- src/main/resources/sonia/scm/jira/template/autoclose.mustache
@@ -1,12 +1,0 @@
-Status modified by changeset {{changeset.id}} of repository {{repository.name}}
-
-Author: {{author}}
-{{#committer}}
-  Committer: {{committer}}
-{{/committer}}
-
-{{changeset.description}}
-
-{{#diffUrl}}
-Changes: {{diffUrl}}
-{{/diffUrl}}
