--- src/main/resources/sonia/scm/jira/template/resend.mustache
@@ -1,12 +1,0 @@
-Author: {{author}}
-{{#committer}}
-  Committer: {{committer}}
-{{/committer}}
-
-{{changeset.description}}
-
-{{#diffUrl}}
-Changes: {{diffUrl}}
-{{/diffUrl}}
-
-The comment was originally sent at {{created}}
