--- src/main/resources/sonia/scm/jira/template/update.mustache
@@ -1,15 +1,0 @@
-{{! It is very important that the Pre tag is on the same line with no spaces as the following content.
-    If the "monospace" option is checked, jira does a terrible job of applying it if there is whitespace }}
-Author: {{author}}
-{{#committer}}
-Committer: {{committer}}
-{{/committer}}
-
-{{commentWrapPre}}{{#descriptionLine}}
-{{.}}
-{{/descriptionLine}}{{commentWrapPost}}
-
-{{#diffUrl}}
-Changes: {{diffUrl}}
-{{/diffUrl}}
-{{#branches}}Branch(es): {{branches}}{{/branches}} {{#bookmarks}}Bookmark(s): {{bookmarks}}{{/bookmarks}}
